+++ js/viz/constants/sim-status.js
