+++ js/viz/lib/generate-svg.js
