+++ js/viz/lib/parseCharge.js
